chore: refactor dependencies to buildSrc
