+++ app/src/androidTest/java/com/wire/android/ui/MainScreenTest.kt
@@ -1,4 +1,4 @@
-package com.wire.android
+package com.wire.android.ui
 
 import androidx.compose.ui.test.assertIsDisplayed
 import androidx.compose.ui.test.junit4.createComposeRule
